Split common into callbacks and messages
