+++ src/dmv/dispatchers/rollSkill.js
@@ -1,6 +1,6 @@
 import * as classes from "../classes";
 
-import { onElementLoad } from "../../common";
+import { onElementLoad } from "../../callbacks";
 
 export const addRollSkillListeners = (store) => onElementLoad(".skills tr .skill-name", () => ready(store));
