+++ src/dmv/dispatchers/rollSpell.js
@@ -1,6 +1,6 @@
 import * as classes from "../classes";
 
-import { onElementLoad } from "../../common";
+import { onElementLoad } from "../../callbacks";
 
 export const addRollSpellListeners = (store) => onElementLoad(".details-columns tr.spell", () => ready(store));
