+++ src/messages.js
@@ -1,0 +1,3 @@
+export const ENQUEUE_COMMANDS = "ENQUEUE";
+export const DEQUEUE_COMMANDS = "DEQUEUE";
+export const CLEAR_QUEUE = "CLEAR";
